tests/requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
construct==2.8.8
pytest-xdist==2.3.0
pycountry==22.3.5
webvtt-py==0.5.1
ttconv==1.2.0